pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kubiya-workflow-sdk"
version = "0.1.0"
description = "SDK for building and managing Kubiya workflows"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Kubiya Team", email = "support@kubiya.ai"}
]
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
server = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "websockets>=11.0",
    "sse-starlette>=1.6.0",
    "python-multipart>=0.0.6",
]
adk = [
    "google-generativeai>=0.8.0",
    "google-genai>=0.3.0",
    "google-adk>=0.1.0",
    "litellm>=1.55.0",
    "together>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
all = [
    "kubiya-workflow-sdk[dev,server,adk]",
]

[project.urls]
Homepage = "https://github.com/kubiya-ai/workflow-sdk"
Documentation = "https://docs.kubiya.ai/workflow-sdk"
Repository = "https://github.com/kubiya-ai/workflow-sdk"
Issues = "https://github.com/kubiya-ai/workflow-sdk/issues"

[project.scripts]
kubiya-workflow = "kubiya_workflow_sdk.cli:main"

[tool.setuptools]
packages = ["kubiya_workflow_sdk"]

[tool.setuptools.package-data]
kubiya_workflow_sdk = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto" 